Remove Python defaults for optional attributes
Summary:
The recent changes to the behavior of optional values in
constructor args missed removing default values in the
Python code and so did not have the desired effect for
non-standard default values unless the BUCK file explicitly
passed `None`. To address that, this changes the Python
generation to always use `None` as a default value for
optional attributes and to let the Java code substitute the
appropriate defaults.

This changes the output of `buck audit rules` and `buck targets`
since those commands operate on the raw maps returned from the
parser (they probably shouldn't do this, those maps should be an
implementation detail).

Test Plan:
Updated `BuckPyFunctionTest` to check behavior for explicit
default values.

Reviewed By: yiding

fbshipit-source-id: a9148ba