Greatly improve BUCK parser exception handling
Summary:
Previously, the `buck.py` parser of `BUCK` files would
internally catch exceptions, format them as text, and pass the
text up to Java along with the rest of the output.

This had a number of problems, but one big part was that
a refactor which moved around `buck.py` code made it so
filtering the stack trace no longer worked.

I wanted to write tests for this whole thing, but I realized
we could just pass the raw exception and stack trace up to
Java and filter and format it there.

That also let me write some nice unit tests.

In addition, I removed the explicit handling of Watchman exceptions
now that D4127744 includes the command information in the exception.

I removed the janky code to try to fall back to the non-Watchman
glob in this case, as it made builds non-reproducible.

Depends On D4127744

Test Plan:
Unit tests added. `ant java-test -Dtest.class=ProjectBuildFileParserTest`
  Set `project.glob_handler = watchman` in `.buckconfig.local`. Ran
  `./bin/buck build buck`, then `kill -9 [watchman-pid]` during
  build. Confirmed we got Watchman exception with error on console.

Reviewed By: yiding

fbshipit-source-id: 4630126