Change ProjectWatches to come from a set of input paths
Summary:
Handle multiple project watches. The assumption of a single
root and single relative path is not necessarily true.

The main idea is to have Main provide a list of watches, and be able to correctly update and make the right watchman requests.

Test Plan: Update tests and manual testing right now. Need to be more systematic about this.

Reviewed By: marcinkosiba

fbshipit-source-id: a13069a