Rename BuildInfo.METADATA_KEY_FOR_XXX to BuildInfo.MetadataKey.XXX
Summary:
It can be confusing to see a name like `METADATA_KEY_FOR_DEP_FILE` as one might confuse this for actually storing a key for something, rather than this being a metadata key under which something is being stored.

This is a straight IntelliJ refactor, no logic changes.

Test Plan: buck build

Reviewed By: ruibm

fbshipit-source-id: 5370207