remove DEPS env var from genrule
Summary:
This looks like an ancient undocumented feature that's unused.

The way that Genrule decides to include this environment variable is via a
substring check, which seems ripe for false-positives.

https://github.com/facebook/buck/issues/999 is someone hitting this
landmine, he had a string `LINT_DEPS` in his code, which triggered the
inclusion of this environment variable, which promptly failed to exec due
to its extreme size.

Test Plan: Watch CI android/ios builds to see nothing suspicious happens.

Reviewed By: Coneko

fbshipit-source-id: a2a2b72