Print Details link on own line to prevent truncating
Summary: The "Details" link in builds often gets truncated by the terminal for going over width. Put it on a new line so this doesn't happen.

Test Plan: run "buck build" with a webserver port set. observe that "Details" gets printed on a new line and does not get truncated.

Reviewed By: marcinkosiba

fbshipit-source-id: ff36e96