Post javac errors as ConsoleEvent so they are logged
Summary:
Currently, if `JavacStep` encounters an error in
`Javac`, it just prints the error directly to the client's
`stderr`.

This means the error is never logged, so it's hard to
diagnose Java compilation errors by inspecting `buck rage`
logs.

In this diff, I change `JavacStep` to instead post stdout and stderr
as `ConsoleEvent`s if the Java compilation failed.

This means java compilation errors now go to the console as well as
the logs, which is a Good Thing.

Test Plan:
Added new unit tests to cover success and failure cases.
`ant java-test -Dtest.class=JavacStepTest`

Reviewed By: dreiss

fbshipit-source-id: 19c55c3