Fix AbstractExecutionContext.createSubContext() losing process executor
Summary:
Currently, `AbstractExecutionContext.createSubContext()` is
used to override the stdout/stderr streams when compiling Java
code in `JavacStep`.

Unfortunately, today, that method discards the execution context's
`ProcessExecutor` and overwrites it with a new
`DefaultProcessExecutor` in order to override stdout and stderr.

I went to write a test for `JavacStep` and found that the
`FakeProcessExecutor` I passed in was lost due to this issue.

To fix it, I introduce a new method `ProcessExecutor.cloneWithOutputStreams()`
and implement it in `DefaultProcessExecutor` and `FakeProcessExecutor`.

Test Plan: Used in test in subsequent diff.

Reviewed By: aiked

fbshipit-source-id: 03f1e29