Stop passing null to getNativeLinkableDeps
Summary:
I added a null argument as a hack while working on library merging, but
this is a better solution, now that we have Java 8.  NativeLinkables are
only required to implement getNativeLinkable[Exported]Deps with no
arguments, to return all possible deps on any platform.  Implementations
*may* also override the single-argument version to return a different
set of deps for specific platforms.

There is a lot of repetitive code here.  I'm wondering if we should try
to push some of these implementations up the class hierarchy, or into a
trait-like interface.

Test Plan: CI

fbshipit-source-id: 59c7015