Add Cgo support in the Go subsystem.
* First get a CxxBinaryDescription to the Go system;
* Add a cgo tool;
* Add cgo source field;
* Add the a cgo rule dependency to GoCompile.

The cgo support was implemented trying to change the least amount of
code possible.
It relies on the existing GoLibrary system which already split the
compilation in different phase:

* compile
* handle potential asm sources
* build an archive (pack)
* link

Now, before the compilation, cgo is run against the cgo sources. After
the run, C and Go sources are generated.
From the C we generate an archive using CxxBinary and we inject the Go
sources to the existing go sources in the GoCompile rule.

The only limitation for now is that we don't handle external C/Cxx
sources as well as external link. However any C code embedded in a Go
file is compiled properly.