Clarify dep-file possibleDepFileSourcePaths
Summary:
It was not clear from the interface Javadoc how this behaves when not present.
This change also makes it a tiny bit clearer and more efficient in the implementiation
as we don't unnecesarily create a hashset copy and check for existence for something that
would always succeed.

Test Plan: unit tests

Reviewed By: asp2insp

fbshipit-source-id: 3d503e7