WatchedFileHashCache: invalidate paths that are "under" event path
Summary:
This handles the case where the watch event is for a symlink to a
directory, in which case we invalidate all cached paths that are
underneath it.  This fixes a bug where files under
`project.read_only_paths` wouldn't get invalidated when their symlink
changed.

Test Plan: added unittest

Reviewed By: illicitonion

fbshipit-source-id: cda7f71