Deduplicate logic for merging of child contexts.
Summary:
We used to manually merge the fields of a child context that contains collected
info from an `include_def`'d file. It's easy to make mistakes when adding a new
field that has to be merged, since all callsites has to be updated (for example,
environment variable and configs used in implicit includes were getting
ignored).

This diff puts all that logic into the previously created abstract base class.

This diff also adds type annotations in docstrings that are sphinx conformant
and can be picked up by intellij pycharms.

Test Plan: CI

Reviewed By: bhamiltoncx

fbshipit-source-id: 8451bd3