IncludeLog: preprocess c/c++, record metadata about includes
Summary:
When building C/C++ sources we record `dep` files, which essentially is a listing of dependencies, i.e. all the files that go into the build: the c/cpp file, headers, headers-included-by-headers, and so on.

For precompiled header support, we will need the same info and a bit more.  To determine compatibility of a preexisting PCH, we need to 1) record the build flags (like `-O3 -g -fomit-frame-pointer`); and (2) the relevant information about include paths.

The first is used to determine whether the prebuilt code/structures are a match for a subsequent build; prebuilt stuff would have to "agree" in terms of debug info, optimizations, and even `#define`'s used during the build.  That list of flag strings can be hashed and used as part of the rule key (as it is now).

For part (2) it's more tricky.  Of course we could include //all// the build flags, including the path information in (2), e.g. `-I` and `-isystem` and so on, and we're done.  And that would not produce incorrect results.  However the number of combinations of include paths is large and will result in //too many// PCH's, with little reuse.  Empirically this resulted in build times breaking even with builds not involving PCHs, and being even worse in other cases (when there are many rules).

Simply dropping all `-I` and other include path flags would result in a greatly decreased number of PCHs, and in a local informal benchmark, but down build time for a large project quite a bit!  However this is dangerous; imagine that the required system libraries between two builds differ, requiring different `-isystem` options.  One might end up unknowingly mixing two sets of system libraries, probably to catastrophic result.

The trick to making this perform well //and// correctly, is to know which include paths "matter" for the PCH build.  But, doing that in a way that doesn't mean rebuilding the PCH again and comparing the outputs; the goal is to reduce PCH builds.

This solution uses a preprocessor option, available in gcc and (soon, when it's merged upstream) in Clang; the `-dI` option, when used with `-E` to preprocess the C/C++ file, will emit the original `#include` directives.  That, along with the line markers in the preprocessed output will allow us to reconstruct, for example, that an include of `stdio.h` resulted in importing the text of `/example/path/to/glibc/6.7.8/include/stdio.h`.

This information is saved alongside a PCH artifact.  In a later build, if we determine that non-include-path related build flags are a match, the candidate PCH's include metadata -- an `IncludeLog`, added in this diff -- is used to determine compatibility in an additional step.  These include events are "replayed", using the new build's set of include flags.  If we see that indeed a `stdio.h` include resulted in the same exact header path, that's a match.  The while IncludeLog is scanned, replayed, and on the first non-match, rejected.

This diff includes:

* IncludeLog library: the preprocessed code parser that finds events; and [de]serialization methods for include events.
* Preprocess the prefix header, capture the inclusion events and file change activity, to determine include path usage by the PCH build.  This `.ilog` file will be used later in determining existing PCH file compatibility, avoiding extra builds later.
* Unit tests exercising these libs, with pretty good coverage, including simple cases as well as real-world output (clang-produced output from all code in modern releases of `glibc` and `stdlibc++`) which had some surprises in it, forcing me to fixing bugs during development

Test Plan:
* Watching CI
* New tests as mentioned above:  `ant -Dtest.class='IncludeLog*Test' java-test`

Reviewed By: Coneko

fbshipit-source-id: d3ab686