Add convenience methods for constructing TestResults.
Summary:
Add convenience methods that populate default values for TestResults.
This change triggers a ton of lint warnings that (unfortunately) we
can't fix, because this class needs to be able to run in the context
of the JUnitRunner, where annotations like Nullable are not necessarily
available.  For that reason, I moved this commit into a separate one
from the call sites that will use it.

Test Plan: Confirm existing tests still pass.

Reviewed By: yiding

fbshipit-source-id: 1c15fe6