External process resource consumption - deep tracking
Summary:
This adds ability to track external processes' resource consumption recursively. In some cases buck launches an external tool that then in turn launches another process. Or sometimes Buck just uses a wrapper script to launch a tool. The problem with that is that if we only gather resource consumption for that process, all we'll see is the cpu real time while the process was waiting on its children, and all the other metrics such as cpu user, cpu kernel, memory, i/o, will effectivly be 0 or close to zero.

I found no direct API to get the process's children tree, so I had to do it backwards. Each process has its parent process pid so I reconstructed the process tree based on this information. Unfortunately, this requires having the list of all the currently running processes. Obtaining this can be expensive so this should not be enabled by default. This might still be useful when more accurate stats are needed so I added a buck config option that can be enabled when appropriate.

Test Plan:
Resource consumption for `genrule-1068664825996620699.sh` before:

  ProcessResourceConsumption{memResident=1327104, memSize=2524295168, cpuReal=21288, cpuUser=2, cpuSys=5, cpuTotal=7, ioBytesRead=65536, ioBytesWritten=0, ioTotal=65536}

Resource consumption for `` after:

  ProcessResourceConsumption{memResident=935264256, memSize=13569228800, cpuReal=39212, cpuUser=28406, cpuSys=10812, cpuTotal=39218, ioBytesRead=667648, ioBytesWritten=8556544, ioTotal=9224192}

Observe that cpu user and sys roughly add up to cpu real now, whereas before they were almost zero. Observe also that we now see the actual resident memory and i/o.

Reviewed By: beefon

fbshipit-source-id: 06aeeb7