change pch filename suffix to `.h.gch`
Summary:
GCC expects to be given a header filename, and it'll first try a precompiled version of that file if available.  Pass this flag, with fake filename: `-include .../somerule-pch-cxx-6789-abcd.h`, and let gcc do its magic of trying the precompiled version `.../somerule-pch-cxx-6789-abcd.h.gch`.

(In truth we only have the latter file but need to give the fake `.h` file to gcc.  Clang doesn't make us jump through this hoop and lets you actually say `-include-pch .../file.gch`.)

This shouldn't change much on a logical level; it's only in preparation for an upcoming feature which will enable PCH for gcc.

Test Plan:
* Ran local tests, on linux and on osx.

  ant -Dtest.class='*IncludeLog*Test' java-test
  ant -Dtest.class='*Precompil*Test' java-test
  ant -Dtest.class='*Cxx*Test' java-test

* watching CI for all the other tests

Reviewed By: Coneko

fbshipit-source-id: faa874c