rust: make sure build output files are recorded
Summary:
buildableContext.recordArtifact() must be called to make
sure the build output is cached properly. Failing to call it succeeds
silently, in that the file exists, and is available to later build
steps, but it's not included in the cache, so later cache hits will
fail to find it.

Test Plan:
$ rm -rf buck-out/
$ buck build --config cache.mode=dir //rust/project/...
$ rm -rf buck-out/gen/
$ buck build --config cache.mode=dir //rust/project/...

This will do a build and cache the results, remove the build, then
rebuild relying on the cache. Previously this failed because the
cache was incomplete.

Reviewed By: bolinfest

fbshipit-source-id: 824a55b