Minor cleanup of PipelineNodeCache.Cache instances in DaemonicParserState
Summary:
Use qualified name `PipelineNodeCache.Cache` instead of just `Cache` throughout
the file, as `Cache` is confusable with guava's `Cache` APIs, which we also use
in this file.

Rename `DaemonicCache` and `DaemonicRawCache` to `DaemonicCacheView` and
`DaemonicRawCacheView`, as these are stateless views on the
`DaemonicParserState`, and don't cache anything themselves.

Use `LoadingCache` instead of ConcurrentMap with manual loading.

Test Plan: CI

Reviewed By: tophyr

fbshipit-source-id: ac008d8