rust: use params.appendExtraDeps()
Summary:
RustNativeLinkables.addNativeDependencies() was using copyWith(),
but this had the effect of moving all the extraDeps into declaredDeps. I'm not this was a real bug since all the deps were preserved in some form, but pretty ugly (and maybe a bug if extraDeps is handled separately?).

Use BuildRuleParams.appendExtraDeps() to just extend extraDeps.

Test Plan: CI

Reviewed By: andrewjcg

fbshipit-source-id: 5fb85a7