Put temporary test files in unique temp dirs
Summary:
The three tests modified in this diff create files in `/tmp/` instead of
in ephemeral temporary directories.

This could cause a problem if two separate `buck test` runs are done by
different users.

In addition to that, the `cxxWriteArgsToFileCreatesDirectoriesIfNeeded`
test does not clean up the directory it creates, which is even worse
given the tests write to the global `/tmp` directory (discovered this
by running buck as root in a container and then running it as the normal
`twsvcscm` user).

Test Plan:
in one terminal:
```
while ! stat /tmp/unexisting_parent_folder; do sleep 10; done
```
in another terminal:
```
./bin/buck test //test/com/facebook/buck/cxx:cxx
```

after the test run:
```
stat /tmp/unexisting_parent_folder
```

Without the patch this folder will be created during the test run,
and does not get removed after.

With the patch the folder is never created (the temp dir is created
under the current working directory) and all temp dirs are removed
when Buck is done by a JVM hook.

Reviewed By: Coneko

fbshipit-source-id: a146103