Convert AndroidBinary to input-based
Summary: Really straightforward, this just has the computeexopackagedepsabi step write the abi key to a file and use that as an input to AndroidBinary. Previously, for non-exopackage builds, we were essentially disabling the abi key optimization, but input-based keys didn't support a good way of doing that. This adds a simple way to disable it dynamically.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 264cb5e