Reorder symlink exception check in getPathForRelativeExistingPath
Summary:
Since symlink check necessitates another system call, reorder it so that it's
only performed if the path does not exist.

This should be faster in most cases (i.e. when files are not symlinks).

Test Plan:
Run `buck targets --show-output` on a rule with many dependencies in the profiler.
In my particular test, observed that getPathForRelativeExistingPath too
approximately half the time as before (3.3s to 1.5s).

Reviewed By: Coneko

fbshipit-source-id: 55e2136