Add support for different ResultTypes that indicate why a test didn't run.
Summary:
Start using new ResultTypes to distinguish between tests that couldn't
be run because of assumption violations, and tests that weren't run
because the tests were disabled or excluded.  This was added in
fbshipit-source-id: 94f731f, then temporarily reverted until
continuous integration tools learned to parse the new types.
Now that this has been fixed, this restores the contents of that commit.

Test Plan:
Confirm existing unit tests still work and continuous integration tests
continue to work correctly with this change.

Reviewed By: marcinkosiba

fbshipit-source-id: 9cb65e6