Replace Phaser with CountDownLatch in ResourcePoolTest
Summary:
Phaser throws an exception when more than the registered number
of parties arrive at the barrier. It's not clear whether this is
a race in Phaser or incorrect usage (see JDK-7058828).
CountDownLatch doesn't care if more threads arrive than the initial
count, so let's use that instead.

Test Plan: CI

Reviewed By: illicitonion

fbshipit-source-id: 68d03d8