Create abi jars on demand
Summary:
Instead of creating abi jars together with their Java dependency
(which requires the description to use `addToIndex` on the rule
it is going to return so that the abi jar can find it in the
build rule resolver), let rules create them on demand using
`requireRule`. This update all the Java rules that use abi jars
except for `android_prebuilt_aar` because that will require more
complex changes.

Test Plan: CI.

Reviewed By: Coneko

fbshipit-source-id: 336b834