Create android_prebuilt_aar abi jars on demand
Summary:
Instead of creating abi jars together with their Java dependency
(which requires the description to use `addToIndex` on the rule
it is going to return so that the abi jar can find it in the build
rule resolver), let rules create them on demand using
`requireRule`. This is a followup to an earlier diff that makes the
same changes for `android_prebuilt_aar`.

Test Plan: CI.

Reviewed By: Coneko

fbshipit-source-id: 3c6a109