Guarantee message printing
Summary:
Currently, this message is only printed if the SuperConsole wins a race
against the shutdown mechanics. As we only actually try to print the
message on a path where we're about to either bubble an exception up to
the top level, or manually call System.exit, synchronously print out the
error and be happy.

Also, stop doing a second check in run - it's redundant.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: 1bf8dd4