CxxPythonExtension: don't fail when platform C/C++ library isn't set
Summary:
This would previously fail when a python platform didn't have a C/C++
library set in it's definition in `.buckconfig`.

Test Plan:
Changed C/C++ python extension tests to not use a set C/C++ library
by default, and add a test that verifies one is used if set.

Reviewed By: Coneko

fbshipit-source-id: 00b9827