Rename RuleKeyBuilderFactory to RuleKeyFactory
Summary:
This interface is about creating RuleKey, not a RuleKeyBuilder. Implementations internally use RuleKeyBuilder, but this is not what this interface is about. It is confusing to see that RuleKeyBuilderFactory.build() creates a RuleKey whereas the name suggests it will create a RuleKeyBuilder.

This is IntelliJ refactor.

Test Plan: all the unit and integration tests.

Reviewed By: Coneko

fbshipit-source-id: 48d0aff