Forbid multiple rules for the same target
Summary:
Throw an exception if multiple rules are generated for the same
target. This can happen if a rule is created and registered with
`addToIndex` while another rule for the same target is being
created.

The previous check only checked equality, which for most (all?)
rules only compares the rule type and the build target which would
not detect when different rules are generated for the same target.

Test Plan: CI.

Reviewed By: Coneko, illicitonion

fbshipit-source-id: ed33793