Freeze rule resolver after action graph is created
Summary:
Make sure that no new build rules are added to the resolver's index
after the target graph to action graph transformation has finished.
It would be preferable to do this by creating a view of the
resolver that doesn't even have the methods to add new rules, but
unfortunately the resolver is stored in too many places (including
many BuildRules) to make it feasible to replace those with an
immutable view and enforce that no new mutable instances will
be passed to the rules in the future.

Test Plan: New unit tests.

Reviewed By: Coneko, illicitonion

fbshipit-source-id: 701b19e