Remove ExecutionEnvironment.getProperty.
Summary:
ExecutionEnvironment.getProperty is supposed to return the value of
properties of the client environment as opposed to the server environment
when using buckd, but it was never implemented, and I'm not sure it even
makes sense as the client environment is python, so it doesn't have java
system properties. In the meantime let's remove it so it's not misleading.

Test Plan: CI

Reviewed By: beefon

fbshipit-source-id: 91f9fdd