Factor out stub jar creation in `MirrorTest`
Summary:
I'm getting ready to introduce a new way of creating ABI jars, and I'd like
to be able to roll it out while keeping the old way. This commit factors out
the part of the test code that actually creates the stubs, so that I can easily
parameterize this test and run it on both implementations.

Test Plan: Unit tests

Reviewed By: asp2insp

fbshipit-source-id: 2a63ae1