Require created rules to have correct targets
Summary:
Descriptions should always create rules for the build targets that
were requested. That is currently not always the case since some
rules add or remove extra flavors, but at least the rest of the
build target is always equivalent. Make sure we don't regress on
this by adding a check to ensure that the returned build rule's
target is apart from flavors equivalent to the requested target.

Test Plan: CI.

Reviewed By: illicitonion

fbshipit-source-id: 00edb0a