Only add symlinks for whitelisted libs in JNI tests
Summary:
Previously, we would build the full symlink tree, but only add
dependencies on the whitelisted libraries.  This had the convenient
effect that dependencies of those libraries were automatically available
at runtime, so there would not be any linker errors.

However, if you forgot to add a library to your whitelist that was only
used from Java, you'd have the symlink but no dependency.  That could
lead to spurious successes (if someone else depended on the target) or
confusing behavior (if the library was changed, but not automatically
rebuilt).

This diff makes us only symlink the libraries that are actually in the
whitelist.  It might requrie test authors to maintain their whitelist
more carefully, but it should give much more predictable behavior.

Test Plan:
Added an integration test.
Updated the old one to properly list its dependency in the whitelist.

Reviewed By: illicitonion

fbshipit-source-id: febe37b