Prevent an exception when generating intellij projects
Summary:
This exception was observed on an `android_library` target that did not have a `manifest` set, thus, the `manifest` fell back to our `.buckconfig[intellij].default_android_manifest_path` which was a relative path.  This fix ensures that the default android manifest path is an absolute path to avoid this exception.

```
java.lang.IllegalArgumentException: 'other' is different type of Path
	at sun.nio.fs.UnixPath.relativize(UnixPath.java:416)
	at sun.nio.fs.UnixPath.relativize(UnixPath.java:43)
	at com.facebook.buck.jvm.java.intellij.IjProjectTemplateDataPreparer.addAndroidManifestPath(IjProjectTemplateDataPreparer.java:458)
	at com.facebook.buck.jvm.java.intellij.IjProjectTemplateDataPreparer.getAndroidProperties(IjProjectTemplateDataPreparer.java:389)
	at com.facebook.buck.jvm.java.intellij.IjProjectWriter.writeModule(IjProjectWriter.java:121)
	at com.facebook.buck.jvm.java.intellij.IjProjectWriter.write(IjProjectWriter.java:87)
	at com.facebook.buck.jvm.java.intellij.IjProject.write(IjProject.java:
Closes https://github.com/facebook/buck/pull/1042

Reviewed By: styurin

fbshipit-source-id: a02545c