Do not override DefaultRuleKeyFactory.newInstance
Summary: `newInstance` is defined in `ReflectiveRuleKeyFactory` and it essentially calls `newBuilder` that subclasses are supposed to implement. `newInstance` then adds bunch of default stuff that should be there, such as buck version, rule name, rule type, etc. There is no need for a subclass to override `newInstance` only to add their own stuff to the rule key. Yhis can be done in `newBuilder` before returning it. This reduces the amount of entanglement in our rule key factiries.

Test Plan: tests

Reviewed By: Coneko

fbshipit-source-id: 6662b76