remove overflow threshold for Watchman
Summary:
Buck currently has a threshold when it interacts with Watchman. If the number of received events is over
10000 then it will fire up an overflow event, that will invalidate parser caches and action graph cache.
It also means that the whole code has to be re-parsed.

This diff removes this threshold. The reasoning behind the removal is that it's not possible re-parsing the
whole repository is faster than re-parsing only the edited files.

Test Plan: - removed the related unit tests.

Reviewed By: marcinkosiba

fbshipit-source-id: a26d8c4