Disable package boundary checks on subsets of cell
Summary:
Right now disabling package boundary checks is global for a whole cell.
This makes it very hard to actually migrate to turn them on, as
regressions are easy. With this change, a subset of a cell can be marked
bad, and can be incrementally improved.

Test Plan: Added tests

Reviewed By: Coneko

fbshipit-source-id: eba2878