Increase buckd idle killer timeout to 1 day
Summary:
This was last upped from 45 minutes to 2 hours in 2014. Since we trust the
daemon a bit more now, let's bump it up to a day. This will keep the daemon
running after someone goes home and comes back the next day.

Any other suggestion for this value welcome. Making it run-time configurable
will be a bit of work, though.

Test Plan: visualscrutiny

Reviewed By: illicitonion

fbshipit-source-id: 6cc5b2f