Update dx from android tools
Summary:
Pull from
https://android.googlesource.com/platform/tools/base/+/gradle_2.2.0/build-system/builder/src/main/java/com/android

Note that this is not actually the DX trunk.  It's a fork by the Android
Tools team that is used in the Android Gradle plugin.  Among other
things (maybe), it has been modified to support multiple concurrent
in-process instances.  This means we no longer need to maintain our
local modifications for the same purpose.

It also includes recent DX updates to speed up merge speed (similar to
our version, which can now be dropped).

Note that this commit will not work on its own.  A second one follows to
make the modifications necessary to get everything working.

Thanks to kageiit for doing the import.

Test Plan: Tested in next commit.

Reviewed By: asp2insp

fbshipit-source-id: 1b0abfd