Provide access to the Compiler Tree API
Summary:
In order to generate ABIs from source, we?ll need access
to the ASTs of the code being compiled. This is provided
by the Compiler Tree API, an implementation of which is
bundled with `javac`.

This commit provides the `abi/source` target and its corresponding
tests with access to the Compiler Tree API, by including `abi/source`
in Buck?s javac plugin instead of in the main body of Buck.

Test Plan: Use this in follow-on commits

Reviewed By: plamenko

fbshipit-source-id: 1dfd166