Remove dependency on ExecutableFinder from buck
Summary:
Copying `ExecutableFinder` to ideabuck source tree to avoid dependency on non-API part of buck.

Also, this code probably contains some functionality related to buck, but not required by the plugin, I'll clean it in following diffs.

The main purpose of this change is to make ideabuck plugin to be independent of buck source code so it can be moved to a different repo.

Test Plan: Build and run plugin, it should detect buck

Reviewed By: JonShemitz

fbshipit-source-id: c53b278