Rewrite daemon startup and idle killer.
Summary:
Rewrite daemon startup to create and run NGServer explicitly, instead of using
`NGServer.main`. This lets me get access to NGServer before it starts to
configure the idle killer without having to dig through the active NGContext for
the server on first access, which makes the code a bit cleaner.

Rewrite daemon idle killer to be simpler by schedule a kill task after the last
concurrent request finishes, instead of polling at every timeout interval, which
may kill the daemon much sooner. Also made it easy to unit test, and added tests
for the common scenarios.

Adds a small `DaemonKillers` class to hold multiple daemon killers. The
motivation for this is to add another daemon killer that kills the server when
the domain socket is deleted, that will take advantage of the code written here,
and will come in a later diff.

Test Plan:
Unit tests for the idle killer logic.

Manual test by:

1. Reduce DAEMON_SLAYER_TIMEOUT to 1 minute.
2. build and run a buck command (e.g. ant clean default && bin/buck query 'deps(buck)'
3. wait for a minute and see that the daemon is gone (e.g. via `ps aux | grep buckd`)
4. run command continuously, see that daemon doesn't get killed while you are
   running commands once every, say, half a minute.

Reviewed By: Coneko

fbshipit-source-id: f6c07e3