Load buck parser code directly from classpath or jar
Summary:
The old scheme would extract a zip file from the jar into a temporary
directory, and set the python path to that directory.  In addition to this
extra unzipping, the zip action can't be specified in IntelliJ, and a
workaround to use the python files in the source directory was needed to run
tests in IntelliJ.

This diff changes how the parser code is loaded by python. Using python's built
in capability of loading from zip files, I put the python source directly into
the jar file, and add the zip file to the python path. In cases where a jar
file is not used, I put the classpath directory into the python path. This
avoids the zip extraction at the start of the buck invocation, and is fully
compatible with a stock intellij.

Test Plan:
Compile and run with ant should succeed:

  `ant clean default && bin/buck targets --show-output buck`

Compile and run with buck should succeed:

  `buck build buck && buck-out/gen/program/buck.pex targets --show-output buck`

Running an integration test with IntelliJ should succeed.

Reviewed By: styurin

fbshipit-source-id: dda8bd6