Catch correct exception, lower log severity
Summary: The exception that is thrown when a file is missing is NoSuchFileException. Also switch to LOG.warn, if the SDKSettings.plist is missing, it's not the end of the world, we can likely skip the SDK and continue.

Test Plan: Ran a buck build with a deleted SDKSettings.plist before and after

Reviewed By: tophyr, mzlee

fbshipit-source-id: ee9ead5