Supply a SourcePathResolver in BuildContext
Summary:
This will allow getBuildSteps to resolve SourcePaths at Step creation
time.

Then we split SourcePathResolver into two interfaces - one of which is
about detecting dependencies, and the other of which is about
converting SourcePaths to (absolute) Paths.

Then we migrate all uses of SourcePathResolver and ProjectFileSystem
within BuildRules to be done in getBuildSteps (or in the Parser).

Then we cut off BuildRules' access to the filesystem, and party.

Test Plan: CI

Reviewed By: plamenko

fbshipit-source-id: d7d9bc2