SocketLossKiller - kill buckd if domain socket has been deleted or changed
Summary:
If someone deletes the buckd socket (e.g. by `hg purge --all`) or `rm -rf`, the
buckd instance will hang around until the idle killer kills it after several
hours.

This diff introduce a background polling job which checks that the socket
exists, and is the same socket file as when the first command was invoked, and
kills buckd if the file is changed or deleted.  Change detection is need in
case the directory is deleted and a new buckd instance is spawned in between
the polling interval.

Test Plan:
Added unit tests

manual testing by running a command via buckd, deleting the `.buckd` directory,
and seeing the buckd instance dissapear from the process list after a few
seconds.

Reviewed By: illicitonion

fbshipit-source-id: d0a46c0