Introduce `buck build <target> --into <path>`
Summary:
It is fairly common to want to `buck build` a target and then use the output of that target.
Historically, that requires doing some hacky shell scripting like the following:

```
buck build <target>
THE_FILE=`buck targets --show-output <target> | awk '{print $2}'`
cd `buck root`
cp $THE_FILE <path>
```

With this change, the above script can now be replaced with the following:

```
buck build <target> --into <path>
```

The one catch is that not every build rule has a valid output that can be copied to an
arbitrary path and is guaranteed to work. (For example, if it relies on relative symlinks
to other paths in `buck-out`.)

This change adds a new default method to the `BuildRule` interface, `supportsBuckBuildInto()`,
that defaults to `return getPathToOutput() != null;`. Individual rules that have an output file must
override this to `return false` to honor the contract of `buck build --into` if their output file cannot
be safely copied and still do what it's supposed to do.

You can see the one case where I explicitly `return false` is in `PythonInPlaceBinary`.

Test Plan:
Added tests to `BuildCommandIntegrationTest`.
Previewed the docs at http://localhost:9811/command/build.html.

Reviewed By: marcinkosiba

fbshipit-source-id: b8a7b43